restructure and removing of unsued functions
